requirements.txt
numpy==1.21.6
pandas==1.3.5
scikit_learn==1.0.2
streamlit==1.23.1
tensorflow_intel==2.11.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
seaborn==0.12.2